refactor(settings): 外观类即时生效设置拆出 useAppearanceDraft
语言 / GUI 主题 / 编辑器外观属即时生效项，与整体保存事务正交，从
useSettingsDraft 抽到独立 hook：后者回归纯草稿/整体保存职责，外观自带
error 由 SettingsModal 合并展示。行为不变。
